add history table

+++ regedit-backend-service/src/main/java/it/istat/mec/regedit/dto/AddressDto.java
@@ -51,8 +51,8 @@ public class AddressDto {
 	private String kmVal;	
 	private String esponenteVal;	
 	private String localitaVal;		
-	private Integer cdpstrEgon;		
-	private Integer cdpcivEgon;		
+	private Long cdpstrEgon;		
+	private Long cdpcivEgon;		
 	private Integer idFonte;	
 	private Integer stratoIndirizzo;	
 	private Integer idRevisore;
