+++ regedit-backend-service/src/main/java/it/istat/mec/regedit/request/UpdateAddressRequest.java
@@ -40,8 +40,8 @@ public class UpdateAddressRequest implements Serializable {
 	private String kmVal;	
 	private String esponenteVal;	
 	private String localitaVal;		
-	private Integer cdpstrEgon;		
-	private Integer cdpcivEgon;		
+	private Long cdpstrEgon;		
+	private Long cdpcivEgon;		
 	private Integer idFonte;		
 	private Integer idRevisore;	
 	private Integer idSupervisor;
